+++ src/main/java/Server/controller/AlbumSingerController.java
@@ -4,6 +4,7 @@ import Server.model.DAO.APIAccountDAO;
 import Server.model.DAO.AlbumSingerDAO;
 import Server.model.DB.AlbumCategorymusicEntity;
 import Server.model.DB.AlbumSingerEntity;
+import Server.model.DTO.AlbumSingerInDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -20,11 +21,12 @@ public class AlbumSingerController {
             method = RequestMethod.POST,
             produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
-    public ResponseEntity<?> registerActor(@RequestBody AlbumSingerEntity albumSingerEntity) {
-//            if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
-//                return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
-//            }
-      albumSingerEntity = albumSingerDAO.save(albumSingerEntity);
+    public ResponseEntity<?> registerActor(@RequestBody AlbumSingerInDTO albumSingerInDTO) {
+        if (albumSingerInDTO == null || albumSingerInDTO.getApiToken() == null || albumSingerInDTO.getApiToken().isEmpty()
+                || apiAccountDAO.checkToken(albumSingerInDTO.getApiToken()) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
+        AlbumSingerEntity albumSingerEntity = albumSingerDAO.save(albumSingerInDTO.getAlbumSingerEntity());
         return new ResponseEntity<>(albumSingerEntity, HttpStatus.CREATED);
     }
 
@@ -32,18 +34,16 @@ public class AlbumSingerController {
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> deleteActor(@PathVariable("idAlbum") Long idAlbum, @PathVariable("idSinger") Long idSinger) {
-//            if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
-//                return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
-//            }
-        if (albumSingerDAO.getId(idAlbum,idSinger) != null) {
-            for (AlbumSingerEntity item : albumSingerDAO.getId(idAlbum,idSinger)) {
+    public ResponseEntity<?> deleteActor(@RequestBody String apiToken, @PathVariable("idAlbum") Long idAlbum, @PathVariable("idSinger") Long idSinger) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
+        if (albumSingerDAO.getId(idAlbum, idSinger) != null) {
+            for (AlbumSingerEntity item : albumSingerDAO.getId(idAlbum, idSinger)) {
                 Long id = item.getId();
                 albumSingerDAO.delete(id);
             }
             return new ResponseEntity<>("Delete Completed", HttpStatus.OK);
         } else return new ResponseEntity<>("Delte Fail", HttpStatus.BAD_REQUEST);
     }
-
-
 }
